rename command bar files
