+++ src/electron/views/automated-checks/components/command-bar.scss
